renaming UI directory
